gmtisp_src/requirements_self.txt
Django==5.0.7
# radius
responses==0.24.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
openwisp-sphinx-theme==1.0.0
freezegun
#openwisp-utils[qa] @ https://github.com/openwisp/openwisp-utils/tarball/master
pylinkvalidator==0.3
lxml==4.9.3
cssselect==1.1.0
# users
django-redis==5.2.0
django-reversion==5.0.8
django-organizations==2.3.1
django-allauth==0.54.0
django-phonenumber-field==7.2.0
django-sesame==3.2.1
phonenumbers==8.13.26
#openwisp-utils[rest,celery] @ https://github.com/openwisp/openwisp-utils/tarball/master
packaging==23.2
django-compress-staticfiles
# other
psycopg2
celery==5.3.6
dj_rest_auth
drf-yasg
django-extensions==3.2.3
django-filter==23.5
django-admin-autocomplete-filter==0.7.1
django-private-storage==3.1.1
django-ipware==6.0.2
django-model-utils==4.3.1
django-sendsms==0.5
django-environ
drf-link-header-pagination==0.2.0
djangosaml2==1.8.0
jsonfield==3.1.0
pyrad==2.4
swapper==1.3.0
weasyprint==60.1

# plans
django-sequences
django_countries
python-stdnum
django-ordered-model
suds
# django-next_url_mixin
# django-bootstrap3

# plan payment
# django-related-admin
# django-payments
# django-plans-payments
